+++ include/nat64/usr/joold/jool_client.h
@@ -1,6 +1,7 @@
 #ifndef JOOL_CLIENT_H
 #define JOOL_CLIENT_H
 
+#include <linux/types.h>
 #include <netlink/netlink.h>
 #include <netlink/socket.h>
 #include <netlink/msg.h>
@@ -19,6 +20,6 @@
 
 	int set_updated_entries(void *data);
 	int get_updated_entries(void);
-	int jool_client_init(int (*cb)(void *, size_t size));
+	int jool_client_init(int (*cb)(void *, __u16 size));
 
 #endif
